🌧 :: 좋아요 중복 및 작품 존재 예외 처리

+++ src/main/java/com/example/lezhinassignment/domain/work/service/AddLikeService.java
@@ -5,6 +5,9 @@ import com.example.lezhinassignment.domain.user.service.facade.UserFacade;
 import com.example.lezhinassignment.domain.work.entity.Like;
 import com.example.lezhinassignment.domain.work.repository.DisLikeRepository;
 import com.example.lezhinassignment.domain.work.repository.LikeRepository;
+import com.example.lezhinassignment.domain.work.repository.WorkRepository;
+import com.example.lezhinassignment.global.exception.user.AlreadyLikeException;
+import com.example.lezhinassignment.global.exception.work.WorkNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,10 +19,19 @@ public class AddLikeService {
 
     private final LikeRepository likeRepository;
     private final DisLikeRepository disLikeRepository;
+    private final WorkRepository workRepository;
     private final UserFacade userFacade;
 
     public void addLike(Long workId) {
         User user = userFacade.currentUser();
+
+        if(!workRepository.existsById(workId)) {
+            throw WorkNotFoundException.EXCEPTION;
+        }
+
+        if(likeRepository.findByUserIdAndWorkId(user.getId(), workId).isPresent()) {
+            throw AlreadyLikeException.EXCEPTION;
+        }
 
         disLikeRepository.deleteByWorkId(workId);
